[CVNG-1179]: Fix decimals in activity risk score sent to UI  (#19278)
* [CVNG-1179]: Fix decimals in activity risk score sent to UI

* [CVNG-1179]: Fix overall activity risk score sent to UI

* tests

+++ 300-cv-nextgen/src/main/java/io/harness/cvng/dashboard/services/impl/HealthVerificationHeatMapServiceImpl.java
@@ -118,7 +118,8 @@ public class HealthVerificationHeatMapServiceImpl implements HealthVerificationH
 
     heatMaps.forEach(heatMap -> {
       Double risk = heatMap.getRiskScore() * 100;
-      categoryRisks.add(CategoryRisk.builder().category(heatMap.getCategory()).risk(risk).build());
+      categoryRisks.add(
+          CategoryRisk.builder().category(heatMap.getCategory()).risk(Double.valueOf(risk.intValue())).build());
       scoreMap.put(heatMap.getCategory(), heatMap.getRiskScore());
     });
